refactor(structure): Move unrelated files out of public ArFS directory PE-50
