+++ src/community_oracle.ts
