+++ src/contract_oracle.ts
