+++ src/gateway_oracle.ts
